fix(ci): debug strange tag "v" generation
on the last run a strange tag "v" was created in addition to "v2.1.6". All the binaries got then uploaded to that "v" release instead of the "v2.1.6" release. Really hard to debug this issue. The problem seems to be in ci.yml as the ci commit had 2 tags attached "v" and "v2.1.6". Removed the "print version" step now as this might be the reason for this mess...

Signed-off-by: Markus Braun <markus.braun@em.ag>