chore(ci): attempt to fix test reports
all test reports that require write access to the prs/repo through a github token (which seems to be all of them) won't work for fork PRs as they haven't got access to secrets. This is now a workaround to write a test summary without any external action.

chore(deps): added pytest-md and pytest-emoji

for test report summary in gh-actions.

Signed-off-by: Markus Braun <markus.braun@em.ag>