chore(code): switched map to generator expression
flake8 complained about a map with a lambda - a generator expression would be more efficient so I changed the map function call to a generator expression.

chore(vscode): fixed devcontainer debugging.

When using a freshly rebuild devcontainer I noticed that debugging won't work because of a missing __main__ module. I added that module and cleaned up the vscode launch config a little.

Signed-off-by: Markus Braun <markus.braun@em.ag>