chore(pypi): added pypi compatible readme
made the main readme pypi compatible and added it to the pypi package by specifing it in the pyproject.toml.
This commit is done in this branch because it will create a new release with the fix (this is easier to test compared to creating a new branch and then another dummy change-release commit).

fixes #36

Signed-off-by: Markus Braun <markus.braun@em.ag>